+++ api/src/main/java/io/opencensus/common/Duration.java
@@ -22,7 +22,6 @@ import static io.opencensus.common.TimeUtils.MILLIS_PER_SECOND;
 import static io.opencensus.common.TimeUtils.NANOS_PER_MILLI;
 
 import com.google.auto.value.AutoValue;
-import io.opencensus.internal.Utils;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.concurrent.Immutable;
 
@@ -115,11 +114,11 @@ public abstract class Duration implements Comparable<Duration> {
    */
   @Override
   public int compareTo(Duration otherDuration) {
-    int cmp = Utils.compareLongs(getSeconds(), otherDuration.getSeconds());
+    int cmp = TimeUtils.compareLongs(getSeconds(), otherDuration.getSeconds());
     if (cmp != 0) {
       return cmp;
     }
-    return Utils.compareLongs(getNanos(), otherDuration.getNanos());
+    return TimeUtils.compareLongs(getNanos(), otherDuration.getNanos());
   }
 
   Duration() {}
